+++ packages/keystone/lib/Keystone/index.js
@@ -155,6 +155,7 @@ module.exports = class Keystone {
     // Deduping here avoids that problem.
     return [
       ...unique(flatten(this.listsArray.map(list => list.getGqlTypes({ skipAccessControl })))),
+      ...unique(flatten(this.listsArray.map(list => list.getGqlMutationTypes()))),
       `"""NOTE: Can be JSON, or a Boolean/Int/String
           Why not a union? GraphQL doesn't support a union including a scalar
           (https://github.com/facebook/graphql/issues/215)"""
@@ -355,9 +356,10 @@ module.exports = class Keystone {
     // memoizing to avoid requests that hit the same type multiple times.
     // We do it within the request callback so we can resolve it based on the
     // request info ( like who's logged in right now, etc)
-    const getListAccessControlForUser = fastMemoize((listKey, operation) => {
+    const getListAccessControlForUser = fastMemoize((listKey, originalInput, operation) => {
       return validateListAccessControl({
         access: this.lists[listKey].access,
+        originalInput,
         operation,
         authentication: { item: user, listKey: authedListKey },
         listKey,
@@ -365,9 +367,10 @@ module.exports = class Keystone {
     });
 
     const getFieldAccessControlForUser = fastMemoize(
-      (listKey, fieldKey, existingItem, operation) => {
+      (listKey, fieldKey, originalInput, existingItem, operation) => {
         return validateFieldAccessControl({
           access: this.lists[listKey].fieldsByPath[fieldKey].access,
+          originalInput,
           existingItem,
           operation,
           authentication: { item: user, listKey: authedListKey },